+++ src/ipbb/cli/vivado.py
@@ -146,6 +146,9 @@ def impl(ctx, *args, **kwargs):
 
 # ------------------------------------------------------------------------------
 @vivado.command('resource-usage', short_help="Resource usage")
+@click.option('-c', '--cell', 'aCell', default=None, help="Submodule name")
+@click.option('-d', '--depth', 'aDepth', type=int, default=1, help="Hierarchy depth")
+@click.option('-f', '--file', 'aFile', type=click.Path(), default=None, help="Output file")
 @click.pass_obj
 @click.pass_context
 def resource_usage(ctx, *args, **kwargs):